Un buon commit: communication between sensors and room-gateway implemented
This commit implements a simple subscription protocol between a room-gateway and multiple sensors. The sensors search for gateways, select the first found and write their mac address on a predefined characteristic. The gateway listen for writes on this characteristic and keep track of mac addresses. Then tries to read data from sensors every x seconds and if an error occurs for more then y times, then removes the mac address from the set of available sensors addressess